Trait ufmt_write::uWrite

pub trait uWrite {
    type Error;
    pub fn write_str(&mut self, s: &str) -> Result<(), Self::Error>;

    pub fn write_char(&mut self, c: char) -> Result<(), Self::Error> { ... }
}

A collection of methods that are required / used to format a message into a stream.

Associated Types

type Error

The error associated to this writer

Loading content...
Required methods

pub fn write_str(&mut self, s: &str) -> Result<(), Self::Error>

Writes a string slice into this writer, returning whether the write succeeded.

This method can only succeed if the entire string slice was successfully written, and this method will not return until all data has been written or an error occurs.

Loading content...
Provided methods

pub fn write_char(&mut self, c: char) -> Result<(), Self::Error>

Writes a char into this writer, returning whether the write succeeded.

A single char may be encoded as more than one byte. This method can only succeed if the entire byte sequence was successfully written, and this method will not return until all data has been written or an error occurs.

Loading content...